frontend/requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
psycopg2-binary==2.9.1
Pillow==9.0.1
matplotlib==3.8.3
pandas
requests
jinja2